feat: Support installing canisters not in dfx.json (#2473)
# Description
Installing NNS canisters involves operating on canisters not listed in dfx.json.

In two places, the existing code checks for canister data in dfx.json.  It doesn't matter if the data is missing, but if the canister is not listed at all, the cli bails out.  E.g. it checks whether the canister has a remote canister ID; it doesn't need to have one, but if the canister is not listed, we crash.  Instead, we would like to treat this in the same way as the data being absent.

We wish to support the case where one dfx server is shared across multiple projects and explicitly do not want to add entries to the dfx.json of a particular project unless the user chooses to do so.

# How Has This Been Tested?
In the development branch `dfx nns install` this change allows NNS canisters to be installed without listing them in dfx.json.