fix: Only add custom section for wasm-format files (#2482)
Previously we attempted to add a section to all custom canisters with a build task, via walrus. This modifies the check to also make sure the file is actually a wasm module, i.e. starts with `b"\0asm"`. That way, build tasks can instead generate a GZIP archive, or whatever else the replica might support in the future. Fixes #2459 and #2357, fixes SDK-696, implements SDK-627.