feat: ic-nns-install part2 (#2486)
# Description

Developers will need to be able to install NNS canisters locally.

This PR adds a new command to do precisely that:
```
dfx start --clean --background
dfx nns install
```

# How Has This Been Tested?
There is a CI test that verifies that the requested NNS canisters have been installed.  I believe that tests that the NNS canisters behave as expected should live in the ic/nns-dapp and internet-identity repos.