fix: compatibility check error should not block canister install (#2561)
# Description

If an error occurs during parsing Candid or stable signature, dfx will error out. This blocks canister upgrade if the generated .did or .most file has bugs that fails the parser.

The fix is to move the checking code into a separate function, dfx checks if the error comes from the 
the actual checking or the preparation steps, and allow users to skip the checking if parsing fails.

# How Has This Been Tested?

Tested manually, but it's hard to test in CI, as it requires pulling in ic-wasm binary as a dependency.

# Checklist:

- [ ] The title of this PR complies with [Conventional Commits](https://www.conventionalcommits.org/en/v1.0.0/).
- [ ] I have edited the CHANGELOG accordingly.
- [ ] I have made corresponding changes to the documentation.