fix: Correct swap wasm file (#2560)
# Description
Previously the governance canister wasm was being used in place of the swap canister wasm by mistake.

# How Has This Been Tested?
With this fix, and another that deals with ic commit mismatches, I am able to create an SNS, which was previously not possible.

Testing SNS creation is to be done in another repo, as most of the complexity is outside the scope of the sdk.  The error was discovered while preparing such a test.

# Checklist:

- [x] The title of this PR complies with [Conventional Commits](https://www.conventionalcommits.org/en/v1.0.0/).
- [x] I have edited the CHANGELOG accordingly.
- [x] I have made corresponding changes to the documentation.