fix: Use NNS did files that matches the wasms (#2559)
# Description
At present NNS did files are downloaded from the head of the ic repo main branch.  However the canister wasms are from a fixed commit, which may not be compatible with head of master.

So now the did files are pulled from the same fixed commit as the canister wasms.