feat: Add `--mode=auto` to `dfx canister install` (#2570)
`dfx deploy` automatically determines whether to use the `install` or `upgrade` modes depending on the canister state. This provides that same functionality to `dfx canister install`, by passing `auto` for the `--mode`. The default remains `install`. 

Fixes #2013 and SDK-590. This was chosen over the suggested feature, `dfx deploy --no-build`, for better consistency with semantics.