chore: If pushing twice to the same branch, later jobs kill earlier ones (#2573)
# Description
CI sometimes, perhaps even often, has to wait minutes to tens of minutes for runners to be available.  Reducing the total number of CPU minutes we spend is a way of freeing up compute capacity.

If a developer (e.g. myself) pushes a branch, then notices a change that needs to be made and pushes again, the earlier job still runs to completion even though it is no longer of interest.  It would be nice to kill the earlier job when it is superseded.

Github's concurrency provides setting to do just this:

https://docs.github.com/en/actions/using-workflows/workflow-syntax-for-github-actions#example-only-cancel-in-progress-jobs-or-runs-for-the-current-workflow

I have opted for the configuration where each workflow has its own groups, so that a kill command from one workflow doesn't kill others.  This lets us choose for which workflows this policy is applied.  I have applied the concurrency limit to every workflow but we may decide that it is only appropriate for some.

# How Has This Been Tested?
* Feel free to push empty commits to this PR and watch CI: `git commit -m bump --allow-empty`
  * I personally have observed this working.  I pushed, waited until there were about 10 entries in https://github.com/dfinity/sdk/actions , then pushed again.  After about 20 seconds the earlier jobs that had not completed were greyed out.

# Checklist:

- [x] The title of this PR complies with [Conventional Commits](https://www.conventionalcommits.org/en/v1.0.0/).
- [x] I have edited the CHANGELOG accordingly.
- [x] I have made corresponding changes to the documentation.
   - I think there is no corresponding documentation so this is a no-op.