fix: Only kill main process on `dfx stop`
Currently the `dfx stop` command tries to kill every child of the main
`dfx` process. This causes the replica to panic when the
`sandbox_launcher` or `canister_sandbox` processes are killed (since it
never expects them to be killed). Instead the `dfx stop` command should
just kill the main `dfx` process and wait for all the children to stop.

This fixes issue
[RUN-351](https://dfinity.atlassian.net/browse/RUN-351).