fix: Lock SNS version (#2587)
# Description
`dfx sns import` was pulling did files from master, so the did files did not match the installed wasms, causing some commands that used to work to fail.

Now, it was actually useful to be able to pull recent did files, so while setting the default reference to download from to the replica commit, I maintained the ability to pull from master by introducing an environment variable `DFX_IC_REF` that can be used to override the commit.  I have added the same env var to `dfx nns install` and `dfx sns import` so that all the wasms are in sync, with the same commit.