chore: Delete OpenSSL (#2659)
I don't know if we used to use OpenSSL for more in the past, but right now we only use it for sha2, which is capably provided by the `sha2` crate instead. In the meantime every last problem I have had with the build system has been due to OpenSSL's inclusion. This PR removes the last vestiges of OpenSSL along with the build system hacks necessary to build it.