feat: nns usability improvements (#2668)
# Description
The `dfx nns`commands are technically ready to use but can be confusing or hard to use, even though they should be very simple.  The following issues are addressed.

## Local network type
`dfx nns install` needs to be run with particular network settings.  There is an existing check that the network type is "system" and an error message but the error message doesn't seem to be helping; it is not clear or explicit enough, it doesn't give people the information they need to configure the network.

**Change:** Expand the error message with explicit steps the user needs to take.

![Screenshot from 2022-10-08 04-42-03](https://user-images.githubusercontent.com/5982633/194685534-23b23a4e-59c5-422d-b36e-5e62d894be63.png)

## Help is not shown
`dfx nns --help` was not showing subcommands.  This was intentional as the nns subcommands were stubbed in, however now the subcommands have been implemented.

**Change:** Make the subcommands public and tweak their descriptions.

## `--network` option is advertised but unused
No current `nns` commands use the `--network` flag, yet the flag is advertised.

**Change:** Remove the `--network` flag from the `nns` subcommand.

# How Has This Been Tested?




# Checklist:

- [x] The title of this PR complies with [Conventional Commits](https://www.conventionalcommits.org/en/v1.0.0/).
- [x] I have edited the CHANGELOG accordingly.
- [x] I have made corresponding changes to the documentation.
  This is a noop