fix: report an error if a custom canister both defines a wasm URL and has a build step (#2697)
# Description

It's not well-defined what should happen if a custom canister has both a `wasm` field that is a URL (meaning dfx downloads the wasm file) and a `build` field (presumably telling how to build the wasm file), so we disallow the combination for the time being.

Fixes https://dfinity.atlassian.net/browse/SDK-776

# How Has This Been Tested?

Added an e2e test.

# Checklist:

- [x] The title of this PR complies with [Conventional Commits](https://www.conventionalcommits.org/en/v1.0.0/).
- [x] I have edited the CHANGELOG accordingly.
- [x] I have made corresponding changes to the documentation.