chore: Change update-frontend-canister.sh to use Docker (#2699)
Rust builds are not reproducible across platforms, even for neutral targets like wasm. #2557 thus unfortunately cannot be run from Macs without using a docker image, and may even have problems across Linux distros. This PR moves the bulk of the build logic into a Dockerfile and has the script use it as a buildkit. 

It also restricts the check to only run when frontend canister code is actually changed, to reduce CI failure spam.

The only change made to the remap logic is that github.com-1ecc6299db9ec823 is no longer shortened to github - if it were not github.com-1ecc6299db9ec823, that'd be extremely important information, because an alternative registry would have been in use. I assume this hash was interpreted as a Git commit hash which could change, but it is a stable string since 2018 edition.