feat: specify effective canister id in provisional_create_canister_with_cycles (#2776)
**Description**

Just like boundary nodes in the mainnet, the reference implementation `ic-ref` now (on master) uses the effective canister id to determine the target subnet of a request. Hence, the effective canister id for the management canister call `provisional_create_canister_with_cycles` must be a valid canister id in the ranges of a subnet on which the canister is created, otherwise the call fails.

When starting `ic-ref`, the subnet configurations are specified using an optional parameter:
```
  --subnet-config ARG      choose initial subnet configurations (default:
                           [(system,"sk1",[(0,1048575)]),(application,"sk2",[(1048576,2097151)])])
```
where each subnet is characterized by its type (e.g., `system` or `application`), a nonce used for generating its key (e.g., `sk1`), and canister ranges of that subnet (specified as a list of pairs of 64-bit integers, the actual canister ID for a 64-bit integer `n` is obtained from the 10-byte binary representation: `n 0x01 0x01`). Hence, the principal with textual representation `rwlgt-iiaaa-aaaaa-aaaaa-cai` is the first canister ID on the system subnet and the principal with textual representation `5v3p4-iyaaa-aaaaa-qaaaa-cai` is the first canister ID on the application subnet in the default `ic-ref` configuration.

The Rust replica implementation ignores the effective canister id for calls to `provisional_create_canister_with_cycles` and thus this change has no effect on the Rust replica implementation.

**In detail, this MR includes the following changes:**
- `ic-hs` is upgraded to master commit `3cc51be5`
- `effective_canister_id` used for `provisional_create_canister_with_cycles` is passed as an command-line argument (defaults to `rwlgt-iiaaa-aaaaa-aaaaa-cai` if not provided or upon parse failure)

**Out-of-scope:**
- override default `--subnet-config` for `ic-ref` (needs further discussion as the low-level input format of the form `[(system,"sk1",[(0,1048575)]),(application,"sk2",[(1048576,2097151)])]` might not be appropriate for `dfx`)