refactor: Box some fields in error types (#2876)
The size of the IdentityError type was approaching the default limits for clippy warnings about error types being too large.  This PR boxes some data on the critical "path", reducing the size (as measured on darwin build) from 96 bytes to 56 bytes, which is less than half of the default limit.