refactor: move wallet canister creation methods to wallet.rs (#2878)
`get_or_create_wallet()` can generate a `DiagnosedError` in some circumstances.  If we were to move this code to the dfx-core crate, I'm not sure how this would fit with concrete error types.  But we don't need to solve that today, because the code we're going to make available in dfx-core doesn't need to create wallet canisters.

This PR moves `get_or_create_wallet()` and the methods that it depends on to a separate file.