refactor: inline Identity::create into IdentityManager::create_new_identity (#2905)
`IdentityManager::create_new_identity()` used to call `Identity::create()`, passing `self`, which in turn called right back into IdentityManager to do a bunch of things.

This PR inlines `Identity::create()` into this single caller.  There are no changes to the moved code.

There may be some opportunity/rationale for refactoring parts of this method back into Identity, but that is not done here.