feat: Get identity from env var if present (#2979)
# Description
`dfx identity use` changes global state.  This means that one script that sets a default identity for its own purposes can interfere with another running concurrently.

This PR proposes a mechanism of scoping a default identity to the current shell.  The identity for a command is:
- as specified by `--identity XXX`, if present, else
- as specified by the environment variable `DFX_IDENTITY=XXX`, if present, else
- as specified by `dfx identity use`
The second of these is scoped to the local shell.