fix: dfx generate requires canister IDs for non-Motoko canisters (#2988)
Previously, non-Motoko canisters required that the canister was created before `dfx generate` could be called.
This requirement is now lifted for all canisters except for canisters of type `"motoko"` or canisters that are listed in (transitive) dependencies of a canister of type `"motoko"`.
It is planned to lift this requirement for Motoko canisters as well, but this requires more work.

Fixes most parts of [SDK-693](https://dfinity.atlassian.net/browse/SDK-693)