fix: remove the loop around the ledger transfer function.
This clarifies to the user what is going on with a sent transaction. The user must know whether a transaction is a duplicate of a past transaction or is a new transaction. using a loop blurrs the lines because once the loop goes around the second time, if it returns a transaction duplicate, the user would not know if it is a duplicate of a past transaction or just a duplicate of the transaction in the first loop.