feat: dfx deploy --by-proposal (#3065)
Adds `dfx deploy --by-proposal`, which uploads ("prepares") asset changes to an asset canister, to be later committed by proposal through the SNS.

Changes HeadersConfig from HashMap to BTreeMap, so that it is sortable, so that `Vec<BatchOperationKind>` can be sortable.

Changes `BatchOperationKind` to be sortable in order to have a deterministic evidence computation.

Fixes https://dfinity.atlassian.net/browse/SDK-969