refactor: methods that read canister info from state tree (#3091)
Upcoming work to support the bitcoin canister needs to determine if a canister exists, what its controllers are, and whether or not it has been installed (has a module hash).

Extracted methods from the `canister info` command for this purpose, while letting the command continue to handle formatting of output.  This turned out to simplify a few other places that also read the module hash.