fix: dfx start will restart replica if it does not report healthy after launch (#3131)
This was almost a refactor, but it can change the behavior in a subtle way.

Previously, if the replica never reported healthy, `dfx start` would display a `panic!`, and not display the Dashboard URL, but would otherwise continue.

Now, if the replica doesn't report healthy within one minute of starting, dfx will terminate the replica process and try again.  Ctrl-C can still abort this process (cause dfx to exit).  I don't expect that this will result in a healthy replica, but who knows?

The motivation for this change: I want to add more "initialization" type steps that happen after launching the replica, such as installing the bitcoin canister.  This will consist of calling one or more `async fn`.  This provides a place to do it.

Further context: I have some work-in-progress that reworks the replica actor to no longer use a separate thread to manage the replica process.  I think it will eventually be nicer (and more correct) than what we have now, but I don't want to hold up the work of updating the replica.