feat: auto-stop canisters on delete (#3135)
Currently, `dfx canister delete` will exit with an error if the canister has not been stopped yet. This PR makes it so that it will confirm deletion with the user instead. (It is intentionally not left friction-free because deletion is destructive; `dfx canister stop` can be viewed as pre-confirming for a particular canister.)

Fixes SDK-893.