fix(ic-asset): consider fewer errors as retryable, and retry create_chunk calls (#3147)
Changes logic to determine whether or not an agent error can be retried during asset synchronization.
From: all errors can be retried, except for certain specific error messages
To: no errors can be retried, except for transport errors and timeout errors

Motivation: I added some new error conditions to the asset canister.  When I tried to verify them in an e2e test, the e2e test appear to hang, due to retries.

While testing this, I discovered that `dfx deploy` would retry a few times but fail quite quickly.  This is because when calling `create_chunk()`, it would retry when waiting for the request to complete, but not while making the call itself.  This PR corrects that as well.