feat: add workflow to update motoko (#3253)
The motivation for this PR is that I'd like to update Motoko.  `scripts/update-motoko.sh` requires nix and niv to be installed.
Since upgrading OSX I no longer have nix and niv installed locally, would like to avoid reinstalling it if I can.

The deleted motoko-updater.yml doesn't work anymore, because it's not possible to run `./scripts/write-dfx-asset-sources.sh` after the niv-updater-action updates nix/sources.json but before it opens a PR.