fix: create_asset fails if asset already exists (#3252)
Makes it so the `CreateAsset` operation and the `create_asset` method fail if the asset already exists.

Previously, it returned without doing anything if the content type matched.

The main motivation for this change is to simplify the description / specification of this operation.

There is no change to the usage: the ic-asset crate's `sync` method already deletes an asset if it will later need to create the same asset, which it must do if the content type changed.

Fixes https://dfinity.atlassian.net/browse/SDK-1134