fix: get_chunk and http_request_streaming_callback require sha256 parameter (#3256)
Make it so `get_chunk()` and `http_request_streaming_callback()` require the sha256 parameter to be `Some(hash)`.

The main motivation is to simplify the documentation/specification for these methods.  It's always been correct to pass these values, which this PR now makes mandatory.

Also added a dfx 0.15.0 migration guide for this and other breaking changes.

Fixes https://dfinity.atlassian.net/browse/SDK-1136