test: re-enable canister http query test (#3266)
This test started failing last month.  It seems that for some reason, canister http outcalls to smartcontracts.org or internetcomputer.org fail with `Could not find a canister id to forward to.`.  While this is interesting, it's not the purpose of this test.

Use `www.githubstatus.com` instead: within the context of a CI run, one query here seems ok.

Addresses https://dfinity.atlassian.net/browse/SDK-1129