docs: add asset canister interface doc (#3258)
Adds documentation for the asset canister interface.  Maybe we will have to move it in order to make it accessible from the portal.

The document doesn't describe any details at all about the `http_request` method.  I'd like to leave that as follow-up work and will open a JIRA ticket to do so if there is agreement.