chore: Clean up use of set_permissions_readwrite (#3303)
Rename set_perms_readwrite to set_permissions_readwrite, move it to dfx_core crate, and use it in places after we copy files.

This is because the source file may not be writable (e.g., installed from system package, or by another user), and we want to make sure the destination file is writable to prevent future writing failures.