feat: Add `use-old-metering` flag to the `dfx` (#3292)
The `use-old-metering` flag enables old metering in replica. The new metering is enabled in the `starter` by default, so this flag is to compare the default new metering with the old one.

The flag is temporary and will be removed in a few months.

Forum post with details: https://forum.dfinity.org/t/new-wasm-instrumentation/22080

Fixes RUN-682