refactor: method-specific error types for IdentityManager::new and initialize (#3309)
Introduces per-method error types for IdentityManager::new() and initialize().

Some of the error enums have `IdentityError` as the error cause.  This happens when a called method returns IdentityError and other methods (that still return IdentityError) call it.  These will change as more methods return their own error types.

The overall direction to look for is removal of variants from the `IdentityError` enum.