fix: remove duplicate builder.with_arg in wallet call (#3372)
When making a call with the wallet, we had two calls to `builder.with_arg()`.  With the current agent this (probably) just adds a parameter that will be ignored, but with the 0.27.0 rust agent this causes a `argument is being set more than once` panic.

This will fix the e2e failures on https://github.com/dfinity/sdk/pull/3370