test: retry mops sources call in playground e2e tests (#3392)
The `mops sources` command in the playground e2e tests fails often.  The retry logic in the JS agent misses some classes of failures (basically, if `fetch` throws).  Pending a fix for that, this change retries the `mops sources` call up to a few times, with simple/naive backoff.

Example test failures:
- https://github.com/dfinity/sdk/actions/runs/6114033428/job/16595080284#step:10:111
- https://github.com/dfinity/sdk/actions/runs/6090058722/job/16524749754#step:10:108
- https://github.com/dfinity/sdk/actions/runs/5929218365/job/16076870508#step:10:109
- https://github.com/dfinity/sdk/actions/runs/5858605218/job/15883717915#step:10:107
- https://github.com/dfinity/sdk/actions/runs/6316833146/job/17152786799?pr=3391