chore: update to icx-proxy-dev built along with rest of replica binaries (#3246)
This updates to the dev version of the icx-proxy, at a commit matching the replica and other released binaries.  It also updates the replica update script to update the icx-proxy in the future.

The new icx-proxy checks certificates.  This has exposed some problems in the asset canister, tracked here and described in the changelog:
- https://dfinity.atlassian.net/browse/SDK-1240
- https://dfinity.atlassian.net/browse/SDK-1245
- https://dfinity.atlassian.net/browse/SDK-1246

So far these appear to be edge cases.  Since they match the behavior that will be seen on mainnet anyway, it seems better to see this behavior in local development as well.

Fixes https://dfinity.atlassian.net/browse/SDK-1200