fix: on PRs, run shellcheck for install scripts (#3469)
Runs the shellcheck workflow for the install script on PRs as well as merge to master, skipping the upload step for PRs.
For the publish-manifest workflow, sets the name to `install-script-shellcheck:required` for later inclusion as a required status.
Does not do this for the dfxvm-install-script-shellcheck, since we'd only remove it again soon, but it will still be reported as a status.