docs: remove description of manual steps from release doc (#3474)
This PR deletes the "MANUAL STEPS" release process description.  It's inaccurate and we have not used it in at least two years.

The only other change is to move the part about whitelisting the frontend canister with the Motoko playground.  There were actually two places that described this, both found in the "MANUAL STEPS" section.