fix: CI e2e:required status could skip even on failure
The always() check provides a "status check"
    https://docs.github.com/en/actions/learn-github-actions/expressions#status-check-functions
If a status check isn't present, a default status check of success() is automatically applied.

This meant that the e2e:required status could be skipped even on test failure,
and when skipped, be marked as successful for purposes of required statuses.

Fixes https://dfinity.atlassian.net/browse/SDK-1474