fix: for dfx new failure if node is installed but npm is not
Would fail with "Failed to scaffold frontend code"

Also, now checks the status code when calling node and npm to
see if they are installed. This means if there is a script
that outputs something like the below, dfx will no longer assume
the program is installed.

$ npm
Command 'npm' not found, but can be installed with:
apt install npm
Please ask your administrator.