fix: Fix capitalization of "Wasm"
This replace incorrect capitalization of "WASM" with "Wasm",
which is the official abbreviation of WebAssembly.

See: https://webassembly.org/

The reason why not all letters are capitalized is because "Wasm"
is not an acronym, but rather an abbreviation.

This change also fixes capitalization of the canister status
output to be consistent wih other fields.