Avoid using git hashes in cargo.toml
This makes it difficult for crates that build on the SDK, since they must also use those exact git hashes on pain of having weird incompatible types. The DFX SNS extension is in this boat. Unfortunately it is difficult for the DFX SNS extension to use that exact git hash since it depends on crates in the IC monorepo which also depend on and use ic-agent.